refactor: move remote-client.ts into activitypub/ (Plan E)
