+++ src/activitypub/remote-client.ts
@@ -13,13 +13,13 @@ import {
   THREAD_MAX_DEPTH,
   THREAD_MAX_REPLIES,
   USER_AGENT,
-} from "./config.js";
-import { type ActivityPubActor, webfingerClient } from "./discovery/webfinger.js";
-import { instanceBlocklist } from "./policy/instance-blocklist.js";
-import { readJsonWithLimit } from "./utils/fetch-helpers.js";
-import { LRUCache } from "./utils/lru-cache.js";
-import { getErrorMessage, validateExternalUrl } from "./utils.js";
-import { DomainSchema } from "./validation/schemas.js";
+} from "../config.js";
+import { type ActivityPubActor, webfingerClient } from "../discovery/webfinger.js";
+import { instanceBlocklist } from "../policy/instance-blocklist.js";
+import { readJsonWithLimit } from "../utils/fetch-helpers.js";
+import { LRUCache } from "../utils/lru-cache.js";
+import { getErrorMessage, validateExternalUrl } from "../utils.js";
+import { DomainSchema } from "../validation/schemas.js";
 
 const logger = getLogger("activitypub-mcp");
