+++ src/mcp/tools-export.ts
@@ -9,8 +9,8 @@ import { getLogger } from "@logtape/logtape";
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { ErrorCode, McpError } from "@modelcontextprotocol/sdk/types.js";
 import { z } from "zod";
+import { remoteClient } from "../activitypub/remote-client.js";
 import { performanceMonitor } from "../performance-monitor.js";
-import { remoteClient } from "../remote-client.js";
 import { validateActorIdentifier, validateDomain } from "../server/index.js";
 import type { RateLimiter } from "../server/rate-limiter.js";
 import { formatErrorWithSuggestion, getErrorMessage, stripHtmlTags } from "../utils.js";
