+++ src/mcp/tools.ts
@@ -9,11 +9,11 @@ import { getLogger } from "@logtape/logtape";
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { ErrorCode, McpError } from "@modelcontextprotocol/sdk/types.js";
 import { z } from "zod";
+import { remoteClient } from "../activitypub/remote-client.js";
 import { dynamicInstanceDiscovery } from "../discovery/dynamic-instance-discovery.js";
 import { instanceDiscovery } from "../discovery/instance-discovery.js";
 import { healthChecker } from "../health-check.js";
 import { performanceMonitor } from "../performance-monitor.js";
-import { remoteClient } from "../remote-client.js";
 import { validateActorIdentifier, validateDomain, validateQuery } from "../server/index.js";
 import type { RateLimiter } from "../server/rate-limiter.js";
 import { formatErrorWithSuggestion, getErrorMessage, stripHtmlTags } from "../utils.js";
