+++ tests/unit/mcp-tools-export.test.ts
@@ -8,7 +8,7 @@ import { registerExportTools } from "../../src/mcp/tools-export.js";
 import { RateLimiter } from "../../src/server/rate-limiter.js";
 
 // Mock dependencies
-vi.mock("../../src/remote-client.js", () => ({
+vi.mock("../../src/activitypub/remote-client.js", () => ({
   remoteClient: {
     fetchActorOutboxPaginated: vi.fn().mockResolvedValue({
       items: [
@@ -75,7 +75,7 @@ vi.mock("@logtape/logtape", () => ({
   }),
 }));
 
-import { remoteClient } from "../../src/remote-client.js";
+import { remoteClient } from "../../src/activitypub/remote-client.js";
 
 describe("MCP Export Tools", () => {
   let mcpServer: McpServer;
